Fix the problem with the text matchers
that deal with needing to add special marks where text is joined across
inline elements and changes context e.g. from numbers to text. There is
a whole slew of text matchers to deal with these special situations.

Now some of these text matchers were not inviked due to a mixup in
priorities. Hopefully this is now straightened out.

To handle this with priorities is very brittle. There should be a better
way.