[mod-sbs] Run all XSpec tests a second time but against an alternative implementation
The alternative implementation first transforms DTBook to EPUB3 HTML,
then translates the EPUB3 HTML to braille, and finally transforms the
EPUB3 HTML back to DTBook.

Most tests currently fail, partly because the braille translation is not
perfect for EPUB3, and partly because the transformation from DTBook to
EPUB3 and back is not perfect.

In order to make this work some changes to epub3-dtbook-migrator were
needed:
- added dtbook-to-epub3.xsl and epub3-to-dtbook.xsl to catalog
- gave all named templates and functions in dtbook-to-epub3.xsl and
  epub3-to-dtbook.xsl a private namespace
- moved some common variables to a shared XSLT