[mod-sbs] XProcSpec tests: pass stylesheet as option instead of linking to it from source
so that when testing via EPUB3, the stylesheet is applied on the EPUB.

Alternative is to leave the link in the DTBook so that it is copied to
the content document of the EPUB, and use the epub3-to-pef option
"apply-document-specific-stylesheets".

In both cases the stylesheet must target both DTBook (for the original
test) and EPUB (for the test via EPUB).