[mod-sbs] Make some table tests work
In order to achieve the desired output we renamed th elements to
td. Note however that we should be able to achieve the same with the old
markup. The preferred solution is to improve the behavior of
"table-header-policy: front" (currently when grouping by row-group the
thead is repeated unnecessarily for tbody and tfoot). An alternative
solution could be to have a special table-header-policy value that makes
th behave like td.

Also changed the CSS so that it is easier to override the default
behavior of tables. For this we're making use of SASS "placeholders".

Also fixed some bugs in css-utils.